Code refactor - merge mixins/draftjs-editing-mixin.js into base/entity-editing-block.js - make infobox editable
