+++ fields/types/html/editor/modifiers/replace-block.js
@@ -4,11 +4,18 @@ import _ from 'lodash';
 import { Entity } from 'draft-js';
 import ENTITY from '../entities';
 
-export function replaceImageBlock(editorState, blockKey, image) {
+export function replaceAtomicBlock(atomicBlockType, editorState, blockKey, value) {
   const content = editorState.getCurrentContent();
   const block = content.getBlockForKey(blockKey);
   const entityKey = block.getEntityAt(0);
-  Entity.mergeData(entityKey, _.pick(image, ENTITY.image.requiredProps));
+  let requiredProps = _.get(ENTITY, [ atomicBlockType, 'requiredProps' ])
+  console.log('replaceAtomicBlock')
+  console.log('value:', value);
+  if (requiredProps) {
+    Entity.mergeData(entityKey, _.pick(value, requiredProps));
+  } else {
+    Entity.mergeData(entityKey, value);
+  }
   return editorState;
 };
